base v1 modif

+++ sakafonao.sql
@@ -8,8 +8,8 @@ use sakafonao;
      password varchar(50),
      nom varchar(50),
      genre varchar(30),
-     taille double,
-     poids double
+     taille int,
+     poids int
  )engine = innoDB;
  INSERT INTO users (id,email,password,nom,genre,taille,poids) VALUES (1,'mana@gmail.com','1234','manasoa','masculin',150,70);
  INSERT INTO users (id,email,password,nom,genre,taille,poids) VALUES (2,'mom@gmail.com','1444','mpiahy','masculin' , 150,72);
